fix(Composite Device): refactors source device behavior into its own struct
